requirements.txt
asgiref==3.8.1
Django==5.1.2
djangorestframework==3.15.2
pillow==10.4.0
python-decouple==3.8
sqlparse==0.5.1
tzdata==2024.2
uuid==1.30
